requirements.txt
streamlit
moviepy
openai-whisper
pathlib
gtts
pygame
googletrans==4.0.0rc1
